requirements.txt
requests
python-dotenv
pandas
matplotlib
networkx
comtradeapicall
tqdm
pytest
world_trade_data
pycountry